GCC/ARMCMx: Remove remaining instances of 'nekkid'
All those naked functions need to put some clothes on:

"The only statements that can be safely included in naked functions
are asm statements that do not have operands. All other statements,
including declarations of local variables, if statements, and so
forth, should be avoided. Naked functions should be used to
implement the body of an assembly function, while allowing the
compiler to construct the requisite function declaration for the
assembler."

See GCC docs.
I've seen that they're not joking about it being unsafe.

Also, displays of public nudity are illegal in most places.